+++ sdk/resourcemanager/recoveryservices/armrecoveryservices/zz_generated_registeredidentities_client.go
@@ -100,7 +100,7 @@ func (client *RegisteredIdentitiesClient) deleteCreateRequest(ctx context.Contex
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-02-01")
+	reqQP.Set("api-version", "2022-03-15-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	return req, nil
 }